requirements.txt
# Core
openai>=1.0
streamlit>=1.30
python-dotenv>=1.0.0

# Utility + validation + data
pydantic>=1.10
python-dateutil>=2.8
pyyaml>=6.0

# Embedding + LLM tooling
tqdm>=4.66
scikit-learn>=1.3
numpy>=1.23

# Logging and debugging
Logging

# Testing
pytest>=7.2

# Optional (future)
#reportlab>=4.0  # For PDF generation